+++ gnl_bonus/get_next_line_bonus.c
@@ -6,7 +6,7 @@
 /*   By: lbardet- <lbardet-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/17 21:22:18 by lbardet-          #+#    #+#             */
-/*   Updated: 2025/10/21 11:22:06 by lbardet-         ###   ########.fr       */
+/*   Updated: 2025/10/22 20:25:38 by lbardet-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,8 @@ char	*ft_strcopy(char *s1, char *s2)
 
 char	*ft_read(char *str, int fd)
 {
-	char	*str1;
-	int		a;
+	char		*str1;
+	ssize_t		a;
 
 	str1 = malloc(BUFFER_SIZE + 1);
 	if (!str1)
@@ -157,16 +157,6 @@ char	*get_next_line(int fd)
 // 		close(fd);
 // 		i++;
 // 	}
-// 	close(fd);
-// 	return (0);
-// }
-
-// int main(void)
-// {
-// 	int fd = open("test.txt", O_RDONLY);
-// 	char *s = get_next_line(fd);
-// 	printf("%s", s);
-// 	free(s);
 // 	close(fd);
 // 	return (0);
 // }
